# mak_it_tz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel",
        "description": "Admin panel for management",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Login",
                "operationId": "auth.login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Logout",
                "operationId": "auth.logout",
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/refresh-token": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Refresh",
                "operationId": "auth.refresh",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "refresh_token": {
                                        "description": "Refresh Token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Register",
                "operationId": "auth.register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "role_id": {
                                        "description": "Id of Role",
                                        "example": "1"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}